+++ db/migrate/20260624210031_add_unit_to_users_and_documents.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class AddUnitToUsersAndDocuments < ActiveRecord::Migration[8.1]
+  def change
+    add_column :users, :enota, :string
+    add_index :users, :enota
+
+    add_column :documents, :unit, :integer, default: 0, null: false
+    add_index :documents, :unit
+  end
+end